headers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,9 +7,9 @@
  */
 int main(void)
 {
-	unsigned long int num1 = 1;
-	unsigned long int num2 = 2;
-	unsigned long int num3;
+	long int num1 = 1;
+	long long int num2 = 2;
+	long long int num3;
 	int count = 0;
 
 	printf("%ld, ", num1);
